Fix potential 64 bit bugs when storing index search results in a 32 bit value.  NSNotFound won't correctly compare.

+++ trunk/src/MGTemplateEngine/ICUTemplateMatcher.m
@@ -151,7 +151,7 @@
 	NSString *argsPattern = @"\"(.*?)(?<!\\\\)\"|'(.*?)(?<!\\\\)'|(\\S+)";
 	NSMutableArray *args = [NSMutableArray array];
 	
-	int location = 0;
+	NSInteger location = 0;
 	while (location != NSNotFound) {
 		NSRange searchRange  = NSMakeRange(location, [argString length] - location);
 		NSRange entireRange = [argString rangeOfRegex:argsPattern options:RKLNoOptions
